install rplidar ROS package from source
As of today, the master branch of the official Slamtec rplidar repository appears 
to work in Noetic when built from source. The most recent commit was made on 
Dec 30, 2019. Since I did not find a binary package, I am adding this to the 
user installation. 

Link to the source I used for testing at the time of this commit: 
https://github.com/Slamtec/rplidar_ros/tree/4f899e670bec2c9e1f26b0969f2de36d23618ef3

Link to the repository:
https://github.com/Slamtec/rplidar_ros