Fix sunshine systemd wrong unit launching
When launching Sunshine via systemd, it uses a version located at
/usr/lib/systemd/user/sunshine.service, which in turn launches
all of the Startup Applications (even the disabled ones) and the
"Welcome to Ubuntu" wizard. This is verified by running:
systemctl --user list-dependencies xdg-desktop-autostart.target
With this commit, we save a version of sunshine.service to
~/.config/systemd/user/sunshine.service and this is the version
used by launching Sunshine via systemd. The startup apps and
wizard are now not appearing.