Set _colcon_cd_root
Using colcon_cd for the first time after opening a new shell prints out the following info to the terminal:

Saved the directory '/home/hello-robot' for future invocations of 'colcon_cd <pkgname>' as well as to return to using 'colcon_cd'. Call 'colcon_cd --reset' to unset the saved path.

This info is unnecessary and confusing. Therefore, we set _colcon_cd_root in .bashrc to avoid having this printout.